chore: simplify story filenames
